requirements.txt
aiogram
python-dotenv
sqlalchemy
asyncpg